# api-spec.yaml
openapi: 3.1.0
info:
  title: Median Code API
  description: Production-ready FastAPI code generation and API entity management
    with Clerk JWT authentication.
  version: 1.0.0
paths:
  /v1/namespaces:
    get:
      tags:
      - Namespaces
      summary: List all namespaces
      description: Retrieve all namespaces accessible to the authenticated user.
      operationId: list_namespaces_v1_namespaces_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NamespaceResponse'
                type: array
                title: Response List Namespaces V1 Namespaces Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Namespaces
      summary: Create a new namespace
      description: Create a new namespace for organizing API entities.
      operationId: create_namespace_v1_namespaces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamespaceCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/namespaces/{namespace_id}:
    get:
      tags:
      - Namespaces
      summary: Get namespace by ID
      description: Retrieve a specific namespace by its ID.
      operationId: get_namespace_v1_namespaces__namespace_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Namespaces
      summary: Update namespace
      description: Update an existing namespace.
      operationId: update_namespace_v1_namespaces__namespace_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamespaceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Namespaces
      summary: Delete namespace
      description: Delete a namespace. Only allowed if the namespace has no entities.
      operationId: delete_namespace_v1_namespaces__namespace_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/fields:
    get:
      tags:
      - Fields
      summary: List all fields
      description: Retrieve all field definitions accessible to the authenticated
        user.
      operationId: list_fields_v1_fields_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldResponse'
                title: Response List Fields V1 Fields Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Fields
      summary: Create a new field
      description: Create a new field definition.
      operationId: create_field_v1_fields_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/fields/{field_id}:
    get:
      tags:
      - Fields
      summary: Get field by ID
      description: Retrieve a specific field by its ID.
      operationId: get_field_v1_fields__field_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          title: Field Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Fields
      summary: Update field
      description: Update an existing field definition.
      operationId: update_field_v1_fields__field_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          title: Field Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Fields
      summary: Delete field
      description: Delete a field. Cannot delete if used in objects.
      operationId: delete_field_v1_fields__field_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          title: Field Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FieldConstraintValueInput:
      properties:
        constraintId:
          type: string
          format: uuid
          title: Constraintid
        value:
          anyOf:
          - type: string
          - type: 'null'
          title: Value
      type: object
      required:
      - constraintId
      title: FieldConstraintValueInput
      description: 'Request schema for attaching a field constraint to a field.


        :ivar constraint_id: Reference to the field constraint definition.

        :ivar value: Parameter value for the constraint (null for parameterless).'
    FieldConstraintValueResponse:
      properties:
        constraintId:
          type: string
          format: uuid
          title: Constraintid
        name:
          type: string
          title: Name
        value:
          anyOf:
          - type: string
          - type: 'null'
          title: Value
      type: object
      required:
      - constraintId
      - name
      title: FieldConstraintValueResponse
      description: 'Response schema for a field constraint value attached to a field.


        :ivar constraint_id: Reference to the field constraint definition.

        :ivar name: Constraint name.

        :ivar value: Parameter value for the constraint.'
    FieldCreate:
      properties:
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000002
        name:
          type: string
          title: Name
          examples:
          - email
        typeId:
          type: string
          format: uuid
          title: Typeid
          examples:
          - 00000000-0000-0000-0001-000000000001
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - User email address
        defaultValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultvalue
          examples:
          - ''
        container:
          anyOf:
          - type: string
            const: List
          - type: 'null'
          title: Container
          examples:
          - List
        constraints:
          items:
            $ref: '#/components/schemas/FieldConstraintValueInput'
          type: array
          title: Constraints
        validators:
          items:
            $ref: '#/components/schemas/FieldValidatorInput'
          type: array
          title: Validators
      type: object
      required:
      - namespaceId
      - name
      - typeId
      title: FieldCreate
      description: 'Request schema for creating a field.


        :ivar namespace_id: Namespace this field belongs to.

        :ivar name: Field name.

        :ivar type_id: Reference to the type definition UUID.

        :ivar description: Field description.

        :ivar default_value: Default value expression.

        :ivar constraints: Constraints to attach to this field.

        :ivar validators: Inline validator definitions for this field.'
    FieldResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          examples:
          - 00000000-0000-0000-0003-000000000001
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000001
        name:
          type: string
          title: Name
          examples:
          - email
        typeId:
          type: string
          format: uuid
          title: Typeid
          examples:
          - 00000000-0000-0000-0001-000000000001
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - User email address
        defaultValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultvalue
          examples:
          - ''
        container:
          anyOf:
          - type: string
            const: List
          - type: 'null'
          title: Container
        usedInApis:
          items:
            type: string
            format: uuid
          type: array
          title: Usedinapis
          examples:
          - - 00000000-0000-0000-0004-000000000001
        constraints:
          items:
            $ref: '#/components/schemas/FieldConstraintValueResponse'
          type: array
          title: Constraints
        validators:
          items:
            $ref: '#/components/schemas/FieldValidatorResponse'
          type: array
          title: Validators
      type: object
      required:
      - id
      - namespaceId
      - name
      - typeId
      title: FieldResponse
      description: 'Response schema for field data.


        :ivar id: Unique identifier for the field.

        :ivar namespace_id: Namespace this field belongs to.

        :ivar name: Field name.

        :ivar type_id: Reference to the type definition UUID.

        :ivar description: Field description.

        :ivar default_value: Default value expression.

        :ivar used_in_apis: Array of API IDs where this field is used.

        :ivar constraints: Constraints attached to this field.

        :ivar validators: Validators attached to this field.'
    FieldUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          examples:
          - updated_field_name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Updated description
        defaultValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultvalue
          examples:
          - new_default
        container:
          anyOf:
          - type: string
            const: List
          - type: 'null'
          title: Container
          examples:
          - List
        constraints:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldConstraintValueInput'
            type: array
          - type: 'null'
          title: Constraints
        validators:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldValidatorInput'
            type: array
          - type: 'null'
          title: Validators
      type: object
      title: FieldUpdate
      description: 'Request schema for updating a field.


        :ivar name: Updated field name.

        :ivar description: Updated description.

        :ivar default_value: Updated default value.

        :ivar constraints: Updated constraints (None = don''t touch, [] = clear all).

        :ivar validators: Updated validators (None = don''t touch, [] = clear all).'
    FieldValidatorInput:
      properties:
        templateId:
          type: string
          format: uuid
          title: Templateid
        parameters:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Parameters
      type: object
      required:
      - templateId
      title: FieldValidatorInput
      description: 'Request schema for attaching a field validator template to a field.


        :ivar template_id: Reference to the field validator template.

        :ivar parameters: Template parameter values.'
    FieldValidatorResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        templateId:
          type: string
          format: uuid
          title: Templateid
        parameters:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Parameters
      type: object
      required:
      - id
      - templateId
      title: FieldValidatorResponse
      description: 'Response schema for an applied field validator.


        :ivar id: Unique identifier for the applied validator.

        :ivar template_id: Reference to the template.

        :ivar parameters: Template parameter values.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NamespaceCreate:
      properties:
        name:
          type: string
          title: Name
          examples:
          - my-api-project
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Custom namespace for my API project
        isDefault:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isdefault
      type: object
      required:
      - name
      title: NamespaceCreate
      description: 'Request schema for creating a namespace.


        :ivar name: Namespace name.

        :ivar description: Optional namespace description.'
    NamespaceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          examples:
          - 00000000-0000-0000-0000-000000000001
        name:
          type: string
          title: Name
          examples:
          - global
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Immutable global templates and examples
        isDefault:
          type: boolean
          title: Isdefault
          examples:
          - false
        locked:
          type: boolean
          title: Locked
          description: True for the provisioned Global namespace (fully read-only).
          readOnly: true
      type: object
      required:
      - id
      - name
      - isDefault
      - locked
      title: NamespaceResponse
      description: 'Response schema for namespace data.


        :ivar id: Unique identifier for the namespace.

        :ivar name: Namespace name.

        :ivar description: Namespace description.

        :ivar is_default: Whether this is the user''s default namespace.'
    NamespaceUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          examples:
          - updated-namespace-name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Updated description
        isDefault:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isdefault
      type: object
      title: NamespaceUpdate
      description: 'Request schema for updating a namespace.


        :ivar name: Updated namespace name.

        :ivar description: Updated namespace description.

        :ivar is_default: Set this namespace as the user''s default.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer